+++ frontend/src/features/signup/SignUp.tsx
@@ -1,16 +1,10 @@
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
-import { z } from 'zod';
 import InputForm from '@/components/InputForm';
 import { Button } from '@/components/ui/button';
 import { Form } from '@/components/ui/form';
-
-const signUpFormSchema = z.object({
-	email: z.email(),
-	password: z.string(),
-});
-
-type SignUpFormType = z.infer<typeof signUpFormSchema>;
+import { type SignUpFormType, signUpFormSchema } from '@/types/signup';
+import useSignUpHook from './useSignUpHook';
 
 function SignUp() {
 	const signUpForm = useForm<SignUpFormType>({
@@ -20,18 +14,8 @@ function SignUp() {
 			password: '',
 		},
 	});
-
-	const onSignUpSubmit = async (data: SignUpFormType) => {
-		console.log(data);
-		const result = await fetch('http://localhost:8000/signup', {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(data),
-		});
-		console.log(result);
-	};
+	
+	const { onSignUpSubmit } = useSignUpHook();
 
 	return (
 		<div className="w-xl mx-auto mt-[60px]">
