+++ frontend/src/types/signup.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const signUpFormSchema = z.object({
+	email: z.email(),
+	password: z.string(),
+});
+
+export type SignUpFormType = z.infer<typeof signUpFormSchema>;
